Move entry storage utils back
