+++ popup/components/EntryDeleteAction.tsx
@@ -4,7 +4,7 @@ import { IconTrash } from "@tabler/icons-react";
 import { useAtomValue } from "jotai";
 
 import { favoriteEntryIdsSetAtom } from "~popup/states/atoms";
-import { deleteEntries } from "~utils/storage/entries";
+import { deleteEntries } from "~utils/storage";
 import { commonActionIconSx } from "~utils/sx";
 
 interface Props {
